feat: finish set_pos_y_of_nodes

+++ rust/src/utils.rs
@@ -10,8 +10,8 @@ mod generate_id;
 pub mod mock_org_chart_data;
 
 pub fn pre_order_traverse_tree<F>(root: Option<Rc<RefCell<Node>>>, mut callback: F)
-where
-    F: FnMut(Rc<RefCell<Node>>) -> (),
+    where
+        F: FnMut(Rc<RefCell<Node>>) -> (),
 {
     if root.is_none() {
         return;
@@ -31,8 +31,8 @@ where
 }
 
 pub fn pre_order_traverse_tree_with_depth<F>(root: Option<Rc<RefCell<Node>>>, mut callback: F)
-where
-    F: FnMut(Rc<RefCell<Node>>, i64) -> (),
+    where
+        F: FnMut(Rc<RefCell<Node>>, usize) -> (),
 {
     if root.is_none() {
         return;
@@ -42,7 +42,7 @@ where
 
     while !queue.is_empty() {
         let (node, depth) = queue.pop_back().unwrap();
-        callback(Rc::clone(&node), depth);
+        callback(Rc::clone(&node), depth as usize);
 
         let children = &node.borrow().children;
         let children_len = children.len();
@@ -54,8 +54,8 @@ where
 }
 
 pub fn post_order_traverse_tree<F>(root: Option<Rc<RefCell<Node>>>, mut callback: F)
-where
-    F: FnMut(Rc<RefCell<Node>>) -> (),
+    where
+        F: FnMut(Rc<RefCell<Node>>) -> (),
 {
     if root.is_none() {
         return;
@@ -81,8 +81,8 @@ where
 }
 
 pub fn bfs_traverse_tree<F>(root: Option<Rc<RefCell<Node>>>, mut callback: F)
-where
-    F: FnMut(Rc<RefCell<Node>>) -> (),
+    where
+        F: FnMut(Rc<RefCell<Node>>) -> (),
 {
     let mut queue = VecDeque::from([root.unwrap()]);
 
@@ -97,8 +97,8 @@ where
 }
 
 pub fn bfs_traverse_tree_with_depth<F>(root: Option<Rc<RefCell<Node>>>, mut callback: F)
-where
-    F: FnMut(Rc<RefCell<Node>>, i64) -> (),
+    where
+        F: FnMut(Rc<RefCell<Node>>, usize) -> (),
 {
     let mut level = -1;
     let mut queue = VecDeque::from([root.unwrap()]);
@@ -109,7 +109,7 @@ where
 
         for _ in 0..len {
             let node = queue.pop_front().unwrap();
-            callback(Rc::clone(&node), level);
+            callback(Rc::clone(&node), level as usize);
 
             for child in &node.borrow().children {
                 queue.push_back(Rc::clone(child));
